backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.11.2
pydantic-settings==2.2.1
email-validator==2.1.0.post1 
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0